replace source packaging with reference to sources
It has been recommended to remove the inclusion of the source from the
produced package. Sources were added to fulfill a license condition for
the CVS application. However, we can alternatively just reference these
sources instead. This commit drops the fetching/packaging of the
source package and introduced a README file describing where sources can
be acquired instead.

Signed-off-by: James Knight <james.d.knight@live.com>